--- src/main/java/seedu/address/commons/util/LambdaUtil/CheckedFunction.java
@@ -1,6 +1,0 @@
-package seedu.address.commons.util.LambdaUtil;
-
-@FunctionalInterface
-public interface CheckedFunction<T, R> {
-    R apply(T t) throws Exception;
-}
